feat(api) - Finalise user session auth
Finalise user session auth for User and Branch controllers

+++ Backend/Controllers/UserController.cs
@@ -3,6 +3,7 @@ using Artifax.DTOs;
 using Artifax.Models;
 using Artifax.Data;
 using Microsoft.EntityFrameworkCore;
+using System.Diagnostics;
 
 namespace Artifax.Controllers
 {
@@ -10,7 +11,9 @@ namespace Artifax.Controllers
     [Route("api/[Controller]")]
     public class UserController : ControllerBase
     {
-        //TODO: Add Session Based Auth
+        const string adminLevel = "Admin";
+        const string employeeLevel = "Employee";
+
         readonly ArtifaxContext context;
 
         public UserController(ArtifaxContext incoming)
@@ -50,24 +53,49 @@ namespace Artifax.Controllers
                 if (_result == null) return NotFound();
                 return AdminReadDto.ToDto(_result);
             }
+
+            [HttpGet("me")]
+            public IActionResult GetCurrentUser()
+            {
+                // Check if a session exists
+                var _userLevel = HttpContext.Session.GetString("UserLevel");
+
+                if (string.IsNullOrEmpty(_userLevel))
+                {
+                    // If no session exists, the user is logged out. 
+                    // Returning 401 Unauthorized tells the frontend to redirect to the login screen.
+                    return Unauthorized("No active session.");
+                }
+
+                // Return an anonymous object (or create a new CurrentUserDto) with the display details
+                return Ok(new 
+                {
+                    UserLevel = _userLevel,
+                    UserEmail = HttpContext.Session.GetString("UserEmail"),
+                    Username = HttpContext.Session.GetString("Username")
+                });
+            }
         #endregion
         
         #region PostRoutes
             [HttpPost("employee")]
             public async Task<ActionResult<EmployeeReadDto>> CreateEmployee (EmployeeWriteDto incoming)
             {
-                //FIXME: Add session Auth
+                //Authentication Test
+                if (HttpContext.Session.GetString("UserLevel") != adminLevel)
+                {
+                    return Unauthorized("Creation attempted by non-admin!");
+                }
 
                 //Check whether employee with incoming details already exists to prevent duplicate emails for login
-                bool _employeeNotFound = await context.Admins.FindAsync(incoming.EmployeeEmail) == null;
-                if (_employeeNotFound) return Unauthorized ("Email Already Exists");
+                bool _employeeFound = await context.Employees.FirstOrDefaultAsync(_employee => _employee.EmployeeEmail == incoming.EmployeeEmail) != null;
+                if (_employeeFound) return Unauthorized ("Email Already Exists");
 
                 //Creates new employee from incoming details
                 Employee _employee = new ()
                 {
-                    BranchId = 0,
+                    BranchId = incoming.BranchID,
                     EmployeeEmail = incoming.EmployeeEmail,
-                    EmployeeId = context.Employees.Max(employee => employee.EmployeeId + 1),
                     EmployeeName = incoming.EmployeeName,
                     EmployeePasswordHash = BCrypt.Net.BCrypt.HashPassword(incoming.EmployeePassword)
                 };
@@ -85,17 +113,21 @@ namespace Artifax.Controllers
             [HttpPost("admin")]
             public async Task<ActionResult<AdminReadDto>> CreateAdmin (AdminWriteDto incoming)
             {
-                //FIXME: Add session Auth
+                //Authentication Test
+                if (HttpContext.Session.GetString("UserLevel") != adminLevel)
+                {
+                    return Unauthorized("Creation attempted by non-admin!");
+                }
 
                 //Check whether admin with incoming details already exists to prevent duplicate emails for login
-                bool _adminNotFound = await context.Admins.FindAsync(incoming.AdminEmail) == null;
-                if (_adminNotFound) return Unauthorized ("Email Already Exists");
+                var _testAdmin = await context.Admins.FirstOrDefaultAsync(admin => admin.AdminEmail == incoming.AdminEmail);
+                
+                if (_testAdmin != null) return Unauthorized ("Email Already Exists");
 
                 //Creates new admin from incoming details
                 Admin _admin = new ()
                 {
                     AdminEmail = incoming.AdminEmail,
-                    AdminId = context.Admins.Count() > 0 ? context.Admins.Max(_i => _i.AdminId + 1) : 0,
                     AdminName = incoming.AdminName,
                     AdminPasswordHash = BCrypt.Net.BCrypt.HashPassword(incoming.AdminPassword)
                 };
@@ -111,38 +143,40 @@ namespace Artifax.Controllers
             }
 
             [HttpPost("employees/login")]
-            public async Task<ActionResult<EmployeeReadDto>> LoginEmployee (string email, string password)
+            public async Task<ActionResult<EmployeeReadDto>> LoginEmployee (LoginDto incoming)
             {
                 //Find employee by email
-                var _employee = await context.Employees.FirstOrDefaultAsync(employee => employee.EmployeeEmail == email);
+                var _employee = await context.Employees.FirstOrDefaultAsync(employee => employee.EmployeeEmail == incoming.Email);
                 if (_employee == null) return Unauthorized ("Employee Not Found");
 
                 //Verify employee password
-                if (!BCrypt.Net.BCrypt.Verify(password, _employee.EmployeePasswordHash))
+                if (!BCrypt.Net.BCrypt.Verify(incoming.Password, _employee.EmployeePasswordHash))
                 {
                     return Unauthorized("Incorrect Password");
                 }
-                HttpContext.Session.SetString("UserLevel","Employee");
-                HttpContext.Session.SetString("UserEmail",email);
+                HttpContext.Session.SetString("UserLevel", employeeLevel);
+                HttpContext.Session.SetString("UserEmail",incoming.Email);
+                HttpContext.Session.SetString("Username", _employee.EmployeeName);
 
                 return Ok(EmployeeReadDto.ToDto(_employee));
             }
 
             [HttpPost("admins/login")]
-            public async Task<ActionResult<AdminReadDto>> LoginAdmin (string email, string password)
+            public async Task<ActionResult<AdminReadDto>> LoginAdmin (LoginDto incoming)
             {
                 //Find admin by email
-                var _admin = await context.Admins.FirstOrDefaultAsync(admin => admin.AdminEmail == email);
+                var _admin = await context.Admins.FirstOrDefaultAsync(admin => admin.AdminEmail == incoming.Email);
                 if (_admin == null) return Unauthorized ("Admin Not Found");
 
                 //Verify admin password
-                if (!BCrypt.Net.BCrypt.Verify(password, _admin.AdminPasswordHash))
+                if (!BCrypt.Net.BCrypt.Verify(incoming.Password, _admin.AdminPasswordHash))
                 {
                     return Unauthorized("Incorrect Password");
                 }
-                HttpContext.Session.SetString("UserLevel","Admin");
-                HttpContext.Session.SetString("UserEmail",email);
-
+                HttpContext.Session.SetString("UserLevel",adminLevel);
+                HttpContext.Session.SetString("UserEmail",incoming.Email);
+                HttpContext.Session.SetString("Username", _admin.AdminName);
+                
                 return Ok(AdminReadDto.ToDto(_admin));
             }
         #endregion
@@ -151,8 +185,12 @@ namespace Artifax.Controllers
             [HttpPatch("employee/branch")]
             public async Task<ActionResult<EmployeeReadDto>> ChangeEmployeeBranch (int EmployeeId, int BranchId)
             {
-                //FIXME: Add session Auth
-
+                //Authentication Check
+                if (HttpContext.Session.GetString("UserLevel") != adminLevel)
+                {
+                    return Unauthorized("Update attempted by non-admin!");
+                }
+                
                 //Check for existing employee
                 var _employee = await context.Employees.FindAsync(EmployeeId);
 
@@ -172,12 +210,18 @@ namespace Artifax.Controllers
             [HttpPatch("employee")]
             public async Task<ActionResult<EmployeeReadDto>> ChangeEmployeeDetails (int id, EmployeeWriteDto incoming)
             {
-                //FIXME: Add session Auth
-
                 //Check for existing employee
                 var _employee = await context.Employees.FindAsync(id);
 
                 if (_employee == null) return NotFound($"Employee with ID: {id.ToString()} could not be found");
+
+                //Authentication Check
+                bool _isAdmin = HttpContext.Session.GetString("UserLevel") == adminLevel;
+
+                if (!_isAdmin)
+                {
+                    return Unauthorized("Update attempted by user with incorrect credentials!");
+                }
 
                 //Update details
                 _employee.EmployeeEmail = incoming.EmployeeEmail;
@@ -192,12 +236,17 @@ namespace Artifax.Controllers
             [HttpPatch("admin")]
             public async Task<ActionResult<EmployeeReadDto>> ChangeAdminDetails (int id, AdminWriteDto incoming)
             {
-                //FIXME: Add session Auth
 
                 //Check for existing admin
                 var _admin = await context.Admins.FindAsync(id);
 
                 if (_admin == null) return NotFound($"Admin with ID: {id.ToString()} could not be found");
+
+                //Authentication Check
+                if (HttpContext.Session.GetString("UserLevel") != adminLevel || HttpContext.Session.GetString("UserEmail") != _admin.AdminEmail)
+                {
+                    return Unauthorized("Update attempted by user with incorrect credentials!");
+                }
 
                 //Update details
                 _admin.AdminEmail = incoming.AdminEmail;
@@ -217,16 +266,30 @@ namespace Artifax.Controllers
                 //Check for existing admin
                 var _admin = await context.Admins.FindAsync(id);
                 if (_admin == null) return NotFound($"Admin with ID: {id.ToString()} could not be found");
+                //Authentication Check
+                if (HttpContext.Session.GetString("UserLevel") != adminLevel)
+                {
+                    return Unauthorized("Delete attempted by non-admin!");
+                }
                 //Remove admin from list
                 context.Admins.Remove(_admin);
                 //Save changes
                 await context.SaveChangesAsync();
+
+                //Clear Session Cookie is Admin deleting self
+                if (HttpContext.Session.GetString("UserEmail") == _admin.AdminEmail) HttpContext.Session.Clear();
+
                 return NoContent();
             }
 
             [HttpDelete("employee")]
             public async Task<IActionResult> DeleteEmployee (int id)
             {
+                //Authentication Check
+                if (HttpContext.Session.GetString("UserLevel") != adminLevel)
+                {
+                    return Unauthorized("Delete attempted by non-admin!");
+                }
                 //Check for existing admin
                 var _employee = await context.Employees.FindAsync(id);
                 if (_employee == null) return NotFound($"Employee with ID: {id.ToString()} could not be found");
